Fix renaming of banner
